bluecap app works with new connect interface

+++ BlueCap/Utils/ConfigStore.swift
@@ -77,17 +77,17 @@ class ConfigStore {
     }
 
     // maximum reconnections
-    class func getMaximumReconnections() -> Int {
+    class func getMaximumReconnections() -> UInt {
         let maximumReconnetions = NSUserDefaults.standardUserDefaults().integerForKey("maximumReconnections")
         if maximumReconnetions == 0 {
             return 5
         } else {
-            return maximumReconnetions
+            return UInt(maximumReconnetions)
         }
     }
     
-    class func setMaximumReconnections(maximumReconnetions:Int) {
-        NSUserDefaults.standardUserDefaults().setInteger(maximumReconnetions, forKey:"maximumReconnections")
+    class func setMaximumReconnections(maximumReconnetions:UInt) {
+        NSUserDefaults.standardUserDefaults().setInteger(Int(maximumReconnetions), forKey:"maximumReconnections")
     }
     
     // scanned services
